Combine the utility of the two servlets now that API is debugged. Works with Cron job call once deployed.
